+++ MichisirubeApplication/app/src/main/java/com/example/michisirubeapplication/ui/naviEvaluation/naviEvaluation.kt
@@ -1,0 +1,26 @@
+package com.example.michisirubeapplication.ui.naviEvaluation
+
+import android.os.Bundle
+import androidx.fragment.app.Fragment
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import com.example.michisirubeapplication.R
+
+class naviEvaluation : Fragment() {
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        arguments?.let {
+
+        }
+    }
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_navi_evaluation, container, false)
+    }
+}
