+++ MichisirubeApplication/app/src/main/res/layout/fragment_navi_evaluation.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".naviEvaluation">
+    tools:context=".ui.naviEvaluation.naviEvaluation">
 
     <!-- TODO: Update blank fragment layout -->
 
@@ -18,7 +18,7 @@
         app:layout_constraintHorizontal_bias="0.0"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:layout_constraintVertical_bias="0.0" />
+        app:layout_constraintVertical_bias="1.0" />
 
     <ImageView
         android:id="@+id/good"
@@ -55,35 +55,44 @@
 
     <TextView
         android:id="@+id/titleEva"
-        android:layout_width="262dp"
-        android:layout_height="22dp"
+        android:layout_width="260dp"
+        android:layout_height="56dp"
         android:layout_marginTop="119dp"
         android:layout_marginBottom="135dp"
-        android:background="#FFFFFF"
         android:text="@string/title_eva"
         android:textColor="@color/brack"
         android:textStyle="bold"
         app:layout_constraintBottom_toTopOf="@+id/good"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintHorizontal_bias="0.47"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintVertical_bias="1.0"
         tools:textSize="18sp" />
 
     <TextView
         android:id="@+id/coments"
-        android:layout_width="262dp"
-        android:layout_height="55dp"
-        android:layout_marginTop="152dp"
-        android:background="#FFFFFF"
+        android:layout_width="298dp"
+        android:layout_height="108dp"
+        android:layout_marginTop="184dp"
         android:text="@string/coments_eva"
-        android:textColor="@color/brack"
+        android:textAllCaps="false"
 
+        android:textColor="@color/brack"
         android:textSize="12sp"
+        android:textStyle="bold"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.496"
+        app:layout_constraintHorizontal_bias="0.513"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
+
+    <Button
+        android:id="@+id/eva_to_home"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        tools:layout_editor_absoluteX="151dp"
+        tools:layout_editor_absoluteY="527dp"
+        tools:text="@string/eva_to_home" />
 
 
 </androidx.constraintlayout.widget.ConstraintLayout>
